refactor(navsidebar): extract types and functions, fix scrollbar style
